Move SendPaymentData to payment module; rename ensure_payment_command -> resume_payment_actor_and_send_command
