+++ crates/fiber-lib/src/fiber/tests/graph.rs
@@ -8,8 +8,8 @@ use crate::fiber::types::{ChannelUpdateChannelFlags, ChannelUpdateMessageFlags, 
 use crate::{
     fiber::{
         graph::{NetworkGraph, RouterHop},
-        network::{get_chain_hash, SendPaymentData},
-        payment::SendPaymentCommand,
+        network::get_chain_hash,
+        payment::{SendPaymentCommand, SendPaymentData},
         types::{ChannelAnnouncement, ChannelUpdate, Hash256, NodeAnnouncement},
     },
     store::Store,
